add subject, delete subject and assign subject APIs

+++ src/main/java/interview/student/repositories/SubjectRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import interview.student.models.Subject;
 
 @Repository
-public interface SubjectRepository extends CrudRepository<Subject, Long> {
+public interface SubjectRepository extends CrudRepository<Subject, Integer> {
 
 }
